Fix the redundant type conversion

+++ tensorrt_llm/_torch/auto_deploy/custom_ops/cuda_backend_causal_conv.py
@@ -186,7 +186,7 @@ def _cuda_cached_causal_conv1d(
         # Use true start offsets for decode tokens (tail after prefills)
         decode_idx = seq_start[num_prefill:].to(torch.long)
         x_decode = inp_flat.index_select(0, decode_idx)  # [num_decode, C_in]
-        slot_idx_decode = slot_idx[num_prefill:].to(torch.long)
+        slot_idx_decode = slot_idx[num_prefill:].to(torch.int32)
         y_dec = causal_conv1d_update(
             x_decode,  # [batch, dim]
             conv_state_cache,
@@ -194,7 +194,7 @@ def _cuda_cached_causal_conv1d(
             bias,
             activation=None,
             cache_seqlens=None,
-            conv_state_indices=slot_idx_decode.to(torch.int32),
+            conv_state_indices=slot_idx_decode,
             pad_slot_id=PAD_SLOT_ID,
         )
